+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Make `AckedTicks` public.
 - Make `NetworkEntityMap` public.
 
 ## [0.9.1] - 2023-08-05